+++ keras_core/random/random_test.py
@@ -216,3 +216,20 @@ class RandomTest(testing.TestCase, parameterized.TestCase):
         self.assertGreater(np.abs(y2 - y3), 1e-4)
 
         seed_generator.global_seed_generator().state.assign(seed)
+
+    def test_shuffle(self):
+        x = np.arange(100).reshape(10, 10)
+
+        # Test axis=0
+        expected = random.shuffle(x, seed=0)
+
+        self.assertFalse(np.all(x == ops.convert_to_numpy(expected)))
+        self.assertAllClose(np.sum(x, axis=0), ops.sum(expected, axis=0))
+        self.assertNotAllClose(np.sum(x, axis=1), ops.sum(expected, axis=1))
+
+        # Test axis=1
+        expected = random.shuffle(x, axis=1, seed=0)
+
+        self.assertFalse(np.all(x == ops.convert_to_numpy(expected)))
+        self.assertAllClose(np.sum(x, axis=1), ops.sum(expected, axis=1))
+        self.assertNotAllClose(np.sum(x, axis=0), ops.sum(expected, axis=0))
